requirements.txt
nextcord
python-dotenv
timeago